+++ tests/parser/test_dataclasses.py
@@ -13,7 +13,7 @@ if PY_VERSION >= (3, 7):
             unit_price: float
             quantity_on_hand: int = 0
 
-        mk_inv, serialize_inv = typeit.type_constructor(InventoryItem)
+        mk_inv, serialize_inv = typeit.TypeConstructor(InventoryItem)
 
         serialized = {
             'name': 'test',