+++ typeit/__init__.py
@@ -3,4 +3,4 @@ from . import sums
 from .combinator.constructor import type_constructor, TypeConstructor
 from .schema.errors import Error
 
-__all__ = ('type_constructor', 'TypeConstructor', 'flags', 'sums', 'Error')
+__all__ = ('TypeConstructor', 'type_constructor', 'flags', 'sums', 'Error')